feat: new import @playwright/test/persistent
This new import exposes the same interface as @playwright/test,
but launches a persistent context for each test instead of
creating a new ephemeral context in a single browser.
